requirements.txt
openai-whisper==20250625
transformers==4.54.0
torch==2.7.1
srt==3.5.3
numpy>=1.21.0
pillow>=8.0.0
